ui: Move TimelineInnerState into its own module
